added new fields parsing
