+++ bot/keyboards.py
@@ -14,15 +14,6 @@ main_menu = InlineKeyboardMarkup(
     )
 
 
-def get_favorite_drugs():
-    """
-    Забирает из БД связь telegram_id и списка препаратов по id.
-    Показывает на кнопках названия препаратов и дозировку
-    select drug.name и drug.dosage
-    """
-    pass
-
-
 # заглушка для ручного тестирования
 favorite_drugs = InlineKeyboardMarkup(
         inline_keyboard=[
